refactor(rag): restructure folder layout and update imports
- Move app/src/ to rag/ package for cleaner imports

- Update all import statements from app.src.* to rag.*

- Rename prompt.py to prompt_builder.py for clarity

- Maintain script accessibility at project root
